Remove authentication-specific test from core to auth (user can change own password in admin backend)
