+++ core/spec/requests/admin/users_spec.rb
@@ -63,19 +63,4 @@ describe "Users" do
       page.should have_content("successfully updated!")
     end
   end
-
-  context "editing own user" do
-    before(:each) do
-      click_link("c@example.com")
-      click_link("Edit")
-    end
-
-    it "should let me edit own password" do
-      fill_in "user_password", :with => "welcome"
-      fill_in "user_password_confirmation", :with => "welcome"
-      click_button "Update"
-
-      page.should have_content("successfully updated!")
-    end
-  end
 end
